- Updated Stock values in ProductTable on NewTransaction Save.

+++ InventoryManagement/Models/Stock.cs
@@ -30,7 +30,7 @@ namespace InventoryManagement.Models
 
         public int ProductID { get; set; }
 
-        public int AvailableQuantity { get; set; }
+        public double AvailableQuantity { get; set; }
 
         public int TotalQuantity { get; set; }
     }
